+++ src/chains/Arbitrum/Common.sol
@@ -134,9 +134,10 @@ abstract contract ArbitrumMixin is
 
             sellToDodoV1(sellToken, bps, dodo, quoteForBase, minBuyAmount);
         } else if (action == uint32(ISettlerActions.RENEGADE.selector)) {
-            (address target, IERC20 baseToken, bytes memory renegadeData) = abi.decode(data, (address, IERC20, bytes));
+            (address target, IERC20 sellToken, bool isSellingBase, bytes memory renegadeData) =
+                abi.decode(data, (address, IERC20, bool, bytes));
 
-            sellToRenegade(target, baseToken, renegadeData);
+            sellToRenegade(target, sellToken, isSellingBase, renegadeData);
         } else {
             return false;
         }